requirements.txt
fastapi
uvicorn
pillow
rembg
onnxruntime
python-multipart
